code/EjercicioGuiado16/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
google-cloud-storage==2.10.0
dicttoxml==1.7.16
werkzeug==2.3.7
flasgger==0.9.7.1
PyMySQL==1.1.0

